Add folder for profile models
